Use git ls-files to determine files in repo
Using `git ls-files` is basically identical to
`git ls-tree --full-tree --name-only -r HEAD` expect for the fact that
ls-tree gives you the full git tree, even when you are in a subfolder.
But since the listing way the FileManager works on the project's top
level and this is where we sniff whether a git repo, this is not an
issue. Hence, we can use `git ls-files --deleted` to exclude the deleted
files from the scan.

Fixes #48