Improve fileName parsing of owasp json bug report
When the a dependency is inside another jar (like uber jars built by
gradle/maven) the filename contains a whitespace between the containing
jar and the depedent jar. The code was assuming that this was always the
case but when the offending jar is in the source tree, the fileName
field contains just the file path.

This caused dependency.fileName.indexOf(' ') to return -1 and the parse
to be incorrect.